Create README.md
White water rafting is a recreation where typically a non-motorized water craft floats and navigates down free flowing rivers. A rafting company organizes and manages trips which most often include experienced guides. The business intent is to provide a thrilling and safe experience to individuals and groups for a fee.